dangm

+++ nmap/services.cc
@@ -557,7 +557,7 @@ void gettoppts(double level, char *portlist, struct scan_lists * ports, double m
         
         ports->prots = NULL;
         
-        for (i = services_by_ratio.begin()++(level); i != services_by_ratio.end(); i++) {
+        for (i = services_by_ratio.begin()++((int)level); i != services_by_ratio.end(); i++) {
             current = &(*i);i.
             if (ptsdata_initialized && !is_port_member(&ptsdata, current))
                 continue;
